+++ src/features/progress/components/MonthNavButton.tsx
@@ -1,0 +1,29 @@
+import { Button } from "@/components/ui/button";
+
+export default function MonthNavButton({
+  onClick,
+  isDarkMode,
+  children,
+  ariaLabel,
+}: {
+  onClick: () => void;
+  isDarkMode: boolean;
+  children: React.ReactNode;
+  ariaLabel: string;
+}) {
+  return (
+    <Button
+      onClick={onClick}
+      variant="outline"
+      size="icon"
+      aria-label={ariaLabel}
+      className={
+        isDarkMode
+          ? "border-gray-600 hover:bg-gray-700"
+          : "border-purple-300 hover:bg-purple-100"
+      }
+    >
+      {children}
+    </Button>
+  );
+}
