+++ src/features/progress/components/ProgressCalendar.tsx
@@ -1,102 +1,20 @@
 import { useState } from "react";
-import { Button } from "@/components/ui/button";
 import { useProgressCalendar } from "@/features/progress/hooks/useProgressCalendar";
+import MonthNavButton from "@/features/progress/components/MonthNavButton";
 import MiscellaneousIcons from "@/icons/miscellaneous";
+import { InsightData } from "@/features/progress/types/InsightData";
+import CalendarDayCircle from "@/features/progress/components/CalendarDayCircle";
 
 const { ChevronLeft, ChevronRight } = MiscellaneousIcons;
-
-interface InsightData {
-  date: string;
-  completionRate: number;
-}
-
-interface ProgressCalendarProps {
-  insightData: InsightData[];
-  isDarkMode: boolean;
-}
-
 const daysOfWeek = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"];
-
-function MonthNavButton({
-  onClick,
-  isDarkMode,
-  children,
-  ariaLabel,
-}: {
-  onClick: () => void;
-  isDarkMode: boolean;
-  children: React.ReactNode;
-  ariaLabel: string;
-}) {
-  return (
-    <Button
-      onClick={onClick}
-      variant="outline"
-      size="icon"
-      aria-label={ariaLabel}
-      className={
-        isDarkMode
-          ? "border-gray-600 hover:bg-gray-700"
-          : "border-purple-300 hover:bg-purple-100"
-      }
-    >
-      {children}
-    </Button>
-  );
-}
-
-function CalendarDayCircle({
-  dayData,
-  isPast,
-  isToday,
-}: {
-  dayData?: InsightData;
-  isPast: boolean;
-  isToday: boolean;
-}) {
-  const showCircle = isPast || isToday;
-  const percent = dayData ? dayData.completionRate : 0;
-
-  return (
-    <div className="relative w-1/2 h-1/2">
-      <svg className="w-full h-full" viewBox="0 0 36 36">
-        <circle
-          className="text-purple-200 dark:text-purple-800"
-          strokeWidth="4"
-          stroke="currentColor"
-          fill="transparent"
-          r="16"
-          cx="18"
-          cy="18"
-        />
-        {showCircle && dayData && (
-          <circle
-            className="text-purple-600 dark:text-purple-400"
-            strokeWidth="4"
-            strokeDasharray={16 * 2 * Math.PI}
-            strokeDashoffset={16 * 2 * Math.PI * (1 - percent / 100)}
-            strokeLinecap="round"
-            stroke="currentColor"
-            fill="transparent"
-            r="16"
-            cx="18"
-            cy="18"
-          />
-        )}
-      </svg>
-      <div className="absolute inset-0 md:flex items-center justify-center hidden">
-        <span className="text-[0.6rem] font-medium text-purple-800 dark:text-purple-200">
-          {percent}%
-        </span>
-      </div>
-    </div>
-  );
-}
 
 export default function ProgressCalendar({
   insightData,
   isDarkMode,
-}: ProgressCalendarProps) {
+}: {
+  insightData: InsightData[];
+  isDarkMode: boolean;
+}) {
   const [currentDate, setCurrentDate] = useState(new Date());
   const { changeMonth, firstDayOfMonth, calendarDays } = useProgressCalendar(
     insightData,
