+++ src/features/progress/types/CalendarDay.ts
@@ -1,0 +1,9 @@
+import { InsightData } from "./InsightData";
+
+export interface CalendarDay {
+  dayOfMonth: number;
+  isPast: boolean;
+  isToday: boolean;
+  date: Date;
+  dayData?: InsightData;
+}
