+++ src/features/progress/types/InsightData.ts
@@ -1,0 +1,4 @@
+export interface InsightData {
+  date: string;
+  completionRate: number;
+}
